requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
PyMuPDF
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai